requirements.txt
pandas==1.3.2
PyYAML==6.0
scikit_learn==1.0.1
torch==1.12.1
